requirements.txt
insightface>=0.7.3
faiss-cpu>=1.7.4
opencv-python>=4.7.0
numpy>=1.23
pillow>=9.0
